Adding <a> support + tolerant parsing
Some HTML nodes are not 3 elements wide and the content would
crash. This is fixed.

The <a> support was added because <link> tags aren't allowed
outside of <head> and a lot of blog writers or CMSes don't
let you easily edit <head> tags for each entry. It is thus
easier to scrape with <a> tags. Both are supported.