Resolving Issue #1, rebar deps lost after reload
No easy way I can think of to test this one, but there is no
regression at all.

The fix works by modifying the rebar.config files found and
injecting lib_dir as the destination path for dependencies.

This in turn created a problem where dependencies were not
tracked from that point on. This prompted the modification
of tend_code_server_jr into tend_code_server, and an additional
call to the lib_dir reloading functions, to make sure the new
deps were added as part of the process.